fixup! feat(suite-native): add experiments manager screen to dev utils
