+++ suite-native/app-init/package.json
@@ -20,6 +20,7 @@
         "@suite-common/wallet-core": "workspace:*",
         "@suite-common/walletconnect": "workspace:*",
         "@suite-native/analytics-redux": "workspace:*",
+        "@suite-native/message-system": "workspace:*",
         "@suite-native/settings": "workspace:*",
         "@suite-native/state": "workspace:*"
     }
