+++ suite-native/message-system/src/index.ts
@@ -1,6 +1,7 @@
 export { ExperimentWrapper } from '@suite-common/message-system';
 
 export * from './messageSystemMiddleware';
+export * from './messageSystemThunks';
 export * from './components/ContextMessage';
 export * from './components/MessageSystemBannerRenderer';
 export * from './components/KillswitchMessageScreen';
